projects/deepseet-knowledge-base/requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
google-generativeai==0.4.1
markdownify==0.11.6
python-dotenv==1.0.1